Remove sort_keys from db format
sort_keys was intended to ensure a consistent order in the database
file, but what ended up happening was a fight between TinyDB (which
wanted to sort the records in numerical ID order) and the JSON library
(which wanted to sort them in string ID order). With Python 3.6, records
and fields are sorted in insertion order (which is reproducible), so
there is no need to TinyDB and JSON to implement their own sorting.